Remove duplicate dependencies
Dependencies had been added to pom.xml besides MANIFEST.MF to enable the
editor to resolve CR and jactr dependencies when they have not been
installed from an Eclipse update site. This interferes with the Maven
build, though, which, unlike the Eclipse editor does not resolve these
dependencies from the Eclipse workspace but attempts to resolve them
from a Maven repository (but no CR and jactr Maven repositories are
configured).

Separate instances of the Eclipse IDE need to be used to develop CR and
jactr on the one hand and jactr-eclipse on the other hand. Dependencies
from CR and jactr need to be installed from an update site in the IDE
for jactr-eclipse.